moved components up
